@@ -3,10 +3,11 @@
 buildscript {
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
+        classpath 'com.android.tools.build:gradle:4.2.2'
+        classpath 'com.vanniktech:gradle-maven-publish-plugin:0.17.0'
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -16,7 +17,7 @@ buildscript {
 allprojects {
     repositories {
         google()
-        jcenter()
+        mavenCentral()
     }
 }
 
@@ -26,25 +27,25 @@ task clean(type: Delete) {
 
 ext {
     minSdkVersion = 14
-    targetSdkVersion = 28
-    compileSdkVersion = 28
-    buildToolsVersion = '28.0.3'
+    targetSdkVersion = 30
+    compileSdkVersion = 30
+    buildToolsVersion = '30.0.3'
     sourceCompatibilityVersion = JavaVersion.VERSION_1_8
     targetCompatibilityVersion = JavaVersion.VERSION_1_8
 }
 
 ext.deps = [
         // Test Dependencies
-        junit             : 'junit:junit:4.12',
+        junit             : 'junit:junit:4.13.2',
         assertjcore       : 'org.assertj:assertj-core:1.7.1',
-        robolectric       : 'org.robolectric:robolectric:3.6.2',
+        robolectric       : 'org.robolectric:robolectric:3.8',
 
         // dependencies
-        appcompatV7       : 'com.android.support:appcompat-v7:28.0.0',
-        androidannotations: 'com.android.support:support-annotations:28.0.0',
-        design            : 'com.android.support:design:28.0.0',
+        appcompat         : 'androidx.appcompat:appcompat:1.3.1',
+        androidannotations: 'androidx.annotation:annotation:1.2.0',
+        mdc               : 'com.google.android.material:material:1.4.0',
 
-        rxandroid         : 'io.reactivex.rxjava2:rxandroid:2.1.0',
-        rxjava            : 'io.reactivex.rxjava2:rxjava:2.2.4',
+        rxandroid         : 'io.reactivex.rxjava2:rxandroid:2.1.1',
+        rxjava            : 'io.reactivex.rxjava2:rxjava:2.2.20',
         timber            : 'com.jakewharton.timber:timber:4.7.1',
 ]
