@@ -7,7 +7,7 @@ public class Constants {
     /**
      * Event type like tags etc...
      */
-    public final class EventType {
+    public static final class EventType {
         public final static String TAG_STORY = "Tom And Jerry";
     }
 }
