@@ -1,14 +1,15 @@
 package com.hwangjr.rxbus.app;
 
 import android.os.Bundle;
-import android.support.design.widget.FloatingActionButton;
-import android.support.design.widget.Snackbar;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
+
+import com.google.android.material.floatingactionbutton.FloatingActionButton;
+import com.google.android.material.snackbar.Snackbar;
 import com.hwangjr.rxbus.RxBus;
 
 import java.util.ArrayList;
@@ -23,11 +24,11 @@ public class MainActivity extends AppCompatActivity {
     /**
      * Mam to birth Tom.
      */
-    private CatMam catMam = new CatMam();
+    private final CatMam catMam = new CatMam();
     /**
      * Mam to birth mouse.
      */
-    private MouseMam mouseMam = new MouseMam();
+    private final MouseMam mouseMam = new MouseMam();
 
     /**
      * Init view and bus provider.
@@ -49,7 +50,7 @@ public class MainActivity extends AppCompatActivity {
                 mouse.squeak();
                 Snackbar.make(view, "Birth a mouse and squeak ! " + mouse, Snackbar.LENGTH_SHORT)
                         .setAction("Action", null).show();
-                Timber.e("Haha, i am " + mouse);
+                Timber.e("Haha, i am %s", mouse);
             }
         });
 
