@@ -1,5 +1,8 @@
 package com.hwangjr.rxbus;
 
+import static junit.framework.Assert.assertEquals;
+import static org.assertj.core.api.Assertions.assertThat;
+
 import com.hwangjr.rxbus.annotation.Produce;
 import com.hwangjr.rxbus.annotation.Subscribe;
 import com.hwangjr.rxbus.thread.EventThread;
@@ -11,9 +14,6 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
-import static junit.framework.Assert.assertEquals;
-import static org.assertj.core.api.Assertions.assertThat;
-
 /**
  * Test that Bus finds the correct producers.
  * <p/>
@@ -21,31 +21,6 @@ import static org.assertj.core.api.Assertions.assertThat;
  */
 @SuppressWarnings("UnusedDeclaration")
 public class AnnotatedProducerFinderTest {
-
-    static class Subscriber {
-        final List<Object> events = new ArrayList<Object>();
-
-        @Subscribe(
-                thread = EventThread.SINGLE
-        )
-        public void subscribe(Object o) {
-            events.add(o);
-        }
-    }
-
-    static class SimpleProducer {
-        static final Object VALUE = new Object();
-
-        int produceCalled = 0;
-
-        @Produce(
-                thread = EventThread.SINGLE
-        )
-        public Object produceIt() {
-            produceCalled += 1;
-            return VALUE;
-        }
-    }
 
     @Test
     public void simpleProducer() {
@@ -70,5 +45,30 @@ public class AnnotatedProducerFinderTest {
         assertThat(producer.produceCalled).isEqualTo(1);
         bus.register(new Subscriber());
         assertThat(producer.produceCalled).isEqualTo(2);
+    }
+
+    static class Subscriber {
+        final List<Object> events = new ArrayList<Object>();
+
+        @Subscribe(
+                thread = EventThread.SINGLE
+        )
+        public void subscribe(Object o) {
+            events.add(o);
+        }
+    }
+
+    static class SimpleProducer {
+        static final Object VALUE = new Object();
+
+        int produceCalled = 0;
+
+        @Produce(
+                thread = EventThread.SINGLE
+        )
+        public Object produceIt() {
+            produceCalled += 1;
+            return VALUE;
+        }
     }
 }
