@@ -1,5 +1,8 @@
 package com.hwangjr.rxbus;
 
+import static junit.framework.Assert.fail;
+import static org.assertj.core.api.Assertions.assertThat;
+
 import com.hwangjr.rxbus.annotation.Subscribe;
 import com.hwangjr.rxbus.thread.EventThread;
 import com.hwangjr.rxbus.thread.ThreadEnforcer;
@@ -15,9 +18,6 @@ import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
-import static junit.framework.Assert.fail;
-import static org.assertj.core.api.Assertions.assertThat;
-
 /**
  * Test that Bus finds the correct subscribers.
  * <p/>
@@ -31,9 +31,9 @@ public class AnnotatedSubscriberFinderTest {
 
     @Ignore // Tests are in extending classes.
     public abstract static class AbstractEventBusTest<S> {
-        abstract S createSubscriber();
-
         private S subscriber;
+
+        abstract S createSubscriber();
 
         S getSubscriber() {
             return subscriber;
@@ -58,6 +58,21 @@ public class AnnotatedSubscriberFinderTest {
      */
     public static class BaseSubscriberFinderTest
             extends AbstractEventBusTest<BaseSubscriberFinderTest.Subscriber> {
+        @Test
+        public void nonSubscriber() {
+            assertThat(getSubscriber().nonSubscriberEvents).isEmpty();
+        }
+
+        @Test
+        public void subscriber() {
+            assertThat(getSubscriber().subscriberEvents).containsExactly(EVENT);
+        }
+
+        @Override
+        Subscriber createSubscriber() {
+            return new Subscriber();
+        }
+
         static class Subscriber {
             final List<Object> nonSubscriberEvents = new ArrayList<Object>();
             final List<Object> subscriberEvents = new ArrayList<Object>();
@@ -73,25 +88,25 @@ public class AnnotatedSubscriberFinderTest {
                 subscriberEvents.add(o);
             }
         }
-
-        @Test
-        public void nonSubscriber() {
-            assertThat(getSubscriber().nonSubscriberEvents).isEmpty();
-        }
-
-        @Test
-        public void subscriber() {
-            assertThat(getSubscriber().subscriberEvents).containsExactly(EVENT);
-        }
-
-        @Override
-        Subscriber createSubscriber() {
-            return new Subscriber();
-        }
     }
 
     public static class AbstractNotAnnotatedInSuperclassTest
             extends AbstractEventBusTest<AbstractNotAnnotatedInSuperclassTest.SubClass> {
+        @Test
+        public void overriddenAndAnnotatedInSubclass() {
+            assertThat(getSubscriber().overriddenAndAnnotatedInSubclassEvents).containsExactly(EVENT);
+        }
+
+        @Test
+        public void overriddenInSubclassNowhereAnnotated() {
+            assertThat(getSubscriber().overriddenInSubclassNowhereAnnotatedEvents).isEmpty();
+        }
+
+        @Override
+        SubClass createSubscriber() {
+            return new SubClass();
+        }
+
         abstract static class SuperClass {
             public abstract void overriddenInSubclassNowhereAnnotated(Object o);
 
@@ -115,10 +130,13 @@ public class AnnotatedSubscriberFinderTest {
                 overriddenAndAnnotatedInSubclassEvents.add(o);
             }
         }
+    }
 
+    public static class NeitherAbstractNorAnnotatedInSuperclassTest
+            extends AbstractEventBusTest<NeitherAbstractNorAnnotatedInSuperclassTest.SubClass> {
         @Test
-        public void overriddenAndAnnotatedInSubclass() {
-            assertThat(getSubscriber().overriddenAndAnnotatedInSubclassEvents).containsExactly(EVENT);
+        public void neitherOverriddenNorAnnotated() {
+            assertThat(getSubscriber().neitherOverriddenNorAnnotatedEvents).isEmpty();
         }
 
         @Test
@@ -126,14 +144,16 @@ public class AnnotatedSubscriberFinderTest {
             assertThat(getSubscriber().overriddenInSubclassNowhereAnnotatedEvents).isEmpty();
         }
 
+        @Test
+        public void overriddenAndAnnotatedInSubclass() {
+            assertThat(getSubscriber().overriddenAndAnnotatedInSubclassEvents).containsExactly(EVENT);
+        }
+
         @Override
         SubClass createSubscriber() {
             return new SubClass();
         }
-    }
 
-    public static class NeitherAbstractNorAnnotatedInSuperclassTest
-            extends AbstractEventBusTest<NeitherAbstractNorAnnotatedInSuperclassTest.SubClass> {
         static class SuperClass {
             final List<Object> neitherOverriddenNorAnnotatedEvents = new ArrayList<Object>();
             final List<Object> overriddenInSubclassNowhereAnnotatedEvents = new ArrayList<Object>();
@@ -166,38 +186,9 @@ public class AnnotatedSubscriberFinderTest {
                 super.overriddenAndAnnotatedInSubclass(o);
             }
         }
-
-        @Test
-        public void neitherOverriddenNorAnnotated() {
-            assertThat(getSubscriber().neitherOverriddenNorAnnotatedEvents).isEmpty();
-        }
-
-        @Test
-        public void overriddenInSubclassNowhereAnnotated() {
-            assertThat(getSubscriber().overriddenInSubclassNowhereAnnotatedEvents).isEmpty();
-        }
-
-        @Test
-        public void overriddenAndAnnotatedInSubclass() {
-            assertThat(getSubscriber().overriddenAndAnnotatedInSubclassEvents).containsExactly(EVENT);
-        }
-
-        @Override
-        SubClass createSubscriber() {
-            return new SubClass();
-        }
     }
 
     public static class FailsOnInterfaceSubscription {
-
-        static class InterfaceSubscriber {
-            @Subscribe(
-                    thread = EventThread.SINGLE
-            )
-            public void whatever(Serializable thingy) {
-                // Do nothing.
-            }
-        }
 
         @Test
         public void subscribingToInterfacesFails() {
@@ -205,6 +196,15 @@ public class AnnotatedSubscriberFinderTest {
                 new Bus(ThreadEnforcer.ANY).register(new InterfaceSubscriber());
                 fail("Annotation finder allowed subscription to illegal interface type.");
             } catch (IllegalArgumentException expected) {
+                // Do nothing.
+            }
+        }
+
+        static class InterfaceSubscriber {
+            @Subscribe(
+                    thread = EventThread.SINGLE
+            )
+            public void whatever(Serializable thingy) {
                 // Do nothing.
             }
         }
